Restore previous indentation at end of IE CC
As Internet Explorer conditional comments may be used to declare
different versions of the same element, unwanted additional indentation
may occur.

This change fixes that by keeping the indentation level on a stack and
restoring the previous level whenever a conditional comment is closed.

A conditional comment should have a matching non-CC implementation,
which will provide the correct indentation for any subsequent lines.