Don't stop on indentation errors unless asked
Most people probably want to get some output rather than an error
message, as in issue #18:

"A minor issue in my usage, the 'Outdented too far' exception. In my
case, I don't really have anything else to do if it's incorrectly
outdented. So I would just ignore it."

This makes the default behaviour to attempt to continue at the current
indentation level if there is an extraneous closing tag or a problem
with an IE conditional comment.

The previous behaviour (raise an error and exit from the command-line
tool) via the :stop_on_errors or --stop-on-errors option.

Closes #18.