Fix self-closing tags regex problems
The regex to allow general self-closing tags was creating problems with
a lot of files. This commit change the order on the mappings to let
work first to the :open_block_element and :open_element mappings. With
this we will use the mapping for general self-closing tags as a last
fallback.

Add failing test to demonstrate bug